Add function `run_in_conda_env` function
This function is a wrapper around `conda run` command.
I located this function in the arcpy module because it comes in handy
when one wants to run a sophisticated code in ArcGis.

ArcGis messes up with a lot of things because it is packed with a custom
conda version and custom conda packages. These packages have strong issues
(conflicts with packages on conda-forge, old versions) and running an
executable in another conda  environment without arcpy is a clever way
to have a reliable and stable execution.

Notable issues avoided by this approach are:
- DLL issues
- Variables issue: Setting environment variables in a python code  may not work
  in ArcGis under some circumstances